requirements.txt
requests
pandas
Flask
urllib3
SQLAlchemy
streamlit
matplotlib
api-offres-emploi
psycopg2